backend/requirement.txt
fastapi
uvicorn
streamlit
psycopg2-binary

langchain-experimental
langchain-openai
openai
pandas
httpx
python-dotenv

faker
requests
pydantic
langchain>=0.1.20

langchain-community

sqlparse

